refactor: all notifications actions api moved to a single notifications controller

+++ Controllers/Notifications.cs
@@ -1,29 +1,46 @@
-using System.Linq;
 using isolaatti_API.isolaatti_lib;
 using isolaatti_API.Models;
 using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 
 namespace isolaatti_API.Controllers
 {
-    [ApiController]
     [Route("/api/[controller]")]
-    public class DeleteNotification : ControllerBase
+    [ApiController]
+    public class Notifications : ControllerBase
     {
         private readonly DbContextApp _db;
 
-        public DeleteNotification(DbContextApp dbContextApp)
+        public Notifications(DbContextApp dbContextApp)
         {
             _db = dbContextApp;
         }
         
         [HttpPost]
-        public IActionResult Index([FromForm] string sessionToken, [FromForm] int notificationId)
+        [Route("GetAllNotifications")]
+        public IActionResult GetAll([FromForm] string sessionToken)
         {
             var accountsManager = new Accounts(_db);
             var user = accountsManager.ValidateToken(sessionToken);
             if (user == null) return Unauthorized("Token is not valid");
             
-            var notification = _db.Notifications.Find(notificationId);
+            var notifications =
+                _db.Notifications
+                    .Where(notification => notification.UserId.Equals(user.Id))
+                    .ToList();
+            return Ok(notifications.OrderByDescending(notification => notification.Id).ToList());
+        }
+
+        [HttpPost]
+        [Route("DeleteNotification")]
+        public IActionResult DeleteANotification([FromForm] string sessionToken, [FromForm] int id)
+        {
+            var accountsManager = new Accounts(_db);
+            var user = accountsManager.ValidateToken(sessionToken);
+            if (user == null) return Unauthorized("Token is not valid");
+            
+            var notification = _db.Notifications.Find(id);
+            if (notification == null) return NotFound("Notification not found");
             if (notification.UserId != user.Id) return Unauthorized("This notification is not yours");
             _db.Notifications.Remove(notification);
             
@@ -31,10 +48,10 @@ namespace isolaatti_API.Controllers
             
             return Ok("Notification deleted successfully");
         }
-
-        [Route("All")]
+        
+        [Route("Delete/All")]
         [HttpPost]
-        public IActionResult DeleteAll([FromForm] string sessionToken, [FromForm] string password)
+        public IActionResult DeleteAll([FromForm] string sessionToken)
         {
             var accountsManager = new Accounts(_db);
             var user = accountsManager.ValidateToken(sessionToken);
@@ -42,8 +59,8 @@ namespace isolaatti_API.Controllers
             
             var notifications = _db.Notifications
                 .Where(notification => notification.UserId.Equals(user.Id));
+            if (!notifications.Any()) return NotFound("No notifications were found, ok");
             _db.Notifications.RemoveRange(notifications);
-            
             _db.SaveChanges();
             return Ok();
         }
